Move filterTypes to types and update imports
- Rename src/components/filterTypes.ts -> src/types/filterTypes.ts and update all imports to the new path (AdvancedFilters, FilterFields, StatFilterBuilder, InventoryGrid). Adjust InventoryGrid imports to use type imports and DEFAULT_FILTER_VALUES from the new location. Also remove the top-level "baseUrl" setting from tsconfig.json. This is a non-functional refactor to centralize type definitions under src/types and keep import paths consistent.
